Initial React/TypeScript commit
