--- .gitignore
@@ -1,5 +1,0 @@
-# Exclude any PSD/AI source
-#*.psd
-#*.ai
-*.psd
-*.ai
