--- dots/dots.css
@@ -1,38 +1,0 @@
-.dot__large {
-  background: center / contain no-repeat url(../images/dots/dotLarge.svg);
-  height: 16px;
-  position: absolute;
-  width: 16px;
-  z-index: -1;
-}
-
-.dot__flashing {
-  animation: flash .5s infinite linear;
-}
-
-@keyframes flash {
-  0% {
-    background: center / contain no-repeat url(../images/dots/dotLarge.svg);
-  }
-  50% {
-    background: black;
-    background-image: none;
-  }
-  100% {
-    background: center / contain no-repeat url(../images/dots/dotLarge.svg);
-  }
-}
-
-.dot__small {
-  background: center / contain no-repeat url(../images/dots/dotSmall.svg);
-  height: 4px;
-  position: absolute;
-  width: 4px;
-  z-index: -1;
-}
-
-.eaten {
-  animation: none;
-  background: black;
-  background-image: none;
-}
