--- index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <link rel="stylesheet" href="./style.css">
-  <link rel="stylesheet" href="./birthday/birthday.css">
-  <link rel="stylesheet" href="./walls/walls.css">
-  <link rel="stylesheet" href="./target/target.css">
-  <link rel="stylesheet" href="./pacman/pacman.css">
-  <link rel="stylesheet" href="./dots/dots.css">
-  <title>Is It Jeffrey's Birthday?</title>
-</head>
-<body>
-  <h1>Is it Jeffrey's Birthday?</h1>
-  <script src="./birthday/birthday.js"></script>
-  <script src="./gameboard/gameboard.js"></script>
-  <script src="./walls/walls.js"></script>
-  <script src="./target/target.js"></script>
-  <script src="./pacman/pacman.js"></script>
-  <script src="./dots/dots.js"></script>
-  <script src="./draw.js"></script>
-</body>
-</html>
