+++ pac-man-bday/images.d.ts
@@ -1,0 +1,3 @@
+declare module '*.svg'
+declare module '*.png'
+declare module '*.jpg'
