+++ pac-man-bday/src/components/App/App.test.tsx
@@ -1,0 +1,13 @@
+/*
+
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import App from './App';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<App />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
+
+*/
