+++ pac-man-bday/src/components/Gameboard/Gameboard.css
@@ -1,0 +1,3 @@
+.Gameboard {
+  color: white;
+}
