+++ pac-man-bday/src/components/Pacman/Pacman.css
@@ -1,0 +1,8 @@
+.Pacman {
+  background: center / contain no-repeat url(../../images/pacman/Pac-Man.svg);
+  color: white;
+  position: absolute;
+  height: 20px;
+  width: 20px;
+  z-index: 2;
+}
