+++ pac-man-bday/src/draw.js
