+++ pac-man-bday/src/index.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import App from './components/App';
+import './index.css';
+import registerServiceWorker from './registerServiceWorker';
+
+ReactDOM.render(
+  <App />,
+  document.getElementById('root') as HTMLElement
+);
+registerServiceWorker();
