--- pacman/pacman.css
@@ -1,6 +1,0 @@
-.pacman {
-  position: absolute;
-  height: 20px;
-  width: 20px;
-  z-index: 2;
-}
